made it display ints instead of floats for health

+++ src/mainGame/HUD.java
@@ -71,7 +71,7 @@ public class HUD {
 		g.drawString("Level: " + level, 15, 75);
 		g.drawString("Extra Lives: " + extraLives, 15, 125);
 		g.drawString("Level Progress: " + levelProgress + "%", 15, 175);
-		g.drawString("Health: " + health + "/" + healthMax, 15, 1050);
+		g.drawString("Health: " + (int)health + "/" + (int)healthMax, 15, 1050);
 		g.drawString("Damage Resist: " +  (2 -player.getDamage()), 15, 275);
 		g.drawString("Player Size: " + player.getPlayerHeight(), 15, 325);
 		g.drawString("Regeneration: " + regen, 15, 375);
